+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - Conditional bug fixed, now works with arrays using deepEquals().
 - Observable.is() bug fixed, now works with arrays using deepEquals().
 - DefaultValueChange.equals() bug fixed, now works with arrays using deepEquals().
+- Value.map() renamed update().
 ### is.codion.swing.common.ui
 - CheckBoxValue removed, related refactoring.
 - DefaultToggleButtonBuilder, AbstractToggleMenuItemBuilder bug fixed, ObservableStateLink added, now sets initial value.